Refactor to pull out REST response types
